backend/pyproject.toml
[project]
name = "inventory-management-backend"
version = "0.1.0"
description = "Inventory management and checkout platform scaffold"
requires-python = ">=3.12,<4.0"
dependencies = [
  "Django>=6.0.5,<7.0.0",
  "djangorestframework>=3.16.1,<4.0.0",
  "djangorestframework-simplejwt>=5.5.1,<6.0.0",
  "drf-spectacular>=0.29.0,<0.30.0",
  "django-cors-headers>=4.9.0,<5.0.0",
  "django-redis>=6.0.0,<7.0.0",
  "redis>=7.0.0,<8.0.0",
  "celery>=5.6.0,<6.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "python-dotenv>=1.2.0,<2.0.0",
  "gunicorn>=25.0.0,<26.0.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
